+++ lib/ooxml_parser/docx_parser/docx_data/document_structure.rb
@@ -1,5 +1,6 @@
 # noinspection RubyInstanceMethodNamingConvention
 require_relative 'document_structure/comment'
+require_relative 'document_structure/comments_extended'
 require_relative 'document_structure/docx_paragraph'
 require_relative 'document_structure/document_background'
 require_relative 'document_structure/document_properties'
@@ -25,6 +26,8 @@ module OoxmlParser
     attr_accessor :theme_colors
     # @return [DocumentSettings] settings
     attr_accessor :settings
+    # @return [CommentsExtended] extended comments
+    attr_accessor :comments_extended
 
     def initialize
       @elements = []
@@ -172,7 +175,8 @@ module OoxmlParser
       end
       OOXMLDocumentObject.xmls_stack.pop
       doc_structure.document_properties = DocumentProperties.new(parent: doc_structure).parse
-      doc_structure.comments = Comment.parse_list
+      doc_structure.comments = Comment.parse_list(parent: doc_structure)
+      doc_structure.comments_extended = CommentsExtended.new(parent: doc_structure).parse
       doc_structure.settings = DocumentSettings.new(parent: doc_structure).parse
       doc_structure
     end