+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/docx_paragraph/docx_paragraph_helper.rb
@@ -1,0 +1,8 @@
+# Stuff for help working with paragraph
+module DocxParagraphHelper
+  # @return [Nil, CommentExtended] extended data for this comment
+  def comment_extend_data
+    return if @paragraph_id.nil?
+    root_object.comments_extended.by_id(@paragraph_id)
+  end
+end